Update config, support full vs minimal metrics.
Group together all config options into a single Config struct with a
global instance.

Add an option --allow_insecure_ssl to allow insecure ssl connections.

Add an option --full_metrics to choose between full and less-full set
of metrics. The minimal set is:
 - CPU: total/user/system, load_average
 - Memory: usage/working set
 - Network: overall (not per-interface)

Things in the full set (and not the minimal set):
 - CPU: per-cpu usage
 - Details on disk I/O utilization
 - Memory: page fault counts
 - Network: per-interface information
 - Filesystem: per-filesystem I/O information
 - task information (sleeping/running/etc)