Add support for container creation/deletion events.
It turns out cAdvisor's events api doesn't return log messages, only
notifications on when containers are created/deleted (and some other
exceptional events like oom, etc). But those are also very useful, so
fetch/forward them as well.

That's done by a new function collectEvents(), which uses the cAdvisor
static API along with time intervals that move forward. The common
task of marshaling structs into data points and sending them to the
jut data node is now handled by sendDataPoiints(). Both collectEvents
and collectMetrics are now run as goroutines so they can't stall each
other. You can enable either events or metrics via the new command
line arguments --events and --metrics.

The event structs have all the information we want to include in
events, with a few small changes (like timestamp -> time, add
source_type: event, etc.) However, one thing that is missing from the
event structs that come back from cAdvisor is the container alias,
which we want to include in jut events. So add some functions to
maintain mappings container name -> container alias in a cache that
can be used by collectEvents, and populate the mappings in
collectMetrics. If a mapping is not found in the cache, do an
on-demand lookup to find the mapping.

In the common case collectMetrics is being run anyway so it's "free"
to populate these mappings. Even if --metrics is false, the function
is run anyway but the points aren't marshaled and sent to the jut data
node. This isn't solely for performance reasons, however--if a
container is deleted, you need to know the mapping ahead of time as an
on-demand fetch wouldn't work as the container has already been
deleted.

Also make a couple of small changes to the build script: add sudo,
change the docker hub account to jut.