PROD-9814 (add docker network metrics) and PROD-9898 (Change default args to jut_cadvisor_agent).
Update to the latest release of google cadvisor when building. This is
only indirectly related to the problem described in PROD-9814 (docker
integration does not capture network traffic metrics), as the true bug
is in cAdvisor itself, and simply upgrading the running version
cadvisor to 0.18.0 allows network metrics to start flowing. But might
as well build against the same version.

To fix PROD-9898, set the flag value for --alsologtostderr to
true, *before* pasing the command line arguments. That way anything
actually specified on the command line still overrides.